Fix enchantments appearing on invalid swords
The new enchantments added by this mod are only intended to work with
their respective swords. Due to an oversight in the vanilla code the
enchantment table only checks compatibility with the enchantment
category and not the enchantment itself. This was later resolved in
Minecraft 1.20.5 however we have to patch in similar functionality
ourselves on older versions.